third_party/WebKit/LayoutTests/presentation/presentation-close-reconnect.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/presentation-service-mock.js"></script>
<button>click me</button>
<script>

async_test(t => {
  presentationServiceMock.then(mockService => {
    const button = document.querySelector('button');
    let connection = null;
    const request = new PresentationRequest('https://example.com');

    waitForClick(_ => {
      request.start().then(conn => {
        connection = conn;
        assert_not_equals(connection, null);
        assert_equals(connection.state, 'connecting');

        connection.onclose = (_ => {
          assert_equals(connection.state, 'closed');
          request.reconnect(connection.id).then(
            t.step_func_done(conn => {
              assert_equals(connection, conn);
              assert_equals(connection.state, 'connecting');
            })
          );
        });
        connection.close();
      });
    }, button);
  });
}, "Test that Presentation.reconnect() resolves with a previously closed presentation connection, whose state is updated to 'connecting.'");

</script>
</body>
</html>
